data_ingestion/kafka/requirements.txt
confluent-kafka
psycopg2-binary
FlightRadarAPI